refactor: move hooks into /lib
