+++ src/components/modals/ClassInfo/ClassInfo.tsx
@@ -10,8 +10,8 @@ import React, { useState } from "react";
 import ClassPopularityMeter from "@/components/schedule/class/ClassPopularityMeter";
 import ClassUsersAvatarGroup from "@/components/schedule/class/ClassUsersAvatarGroup";
 import ConfirmationDialog from "@/components/utils/ConfirmationDialog";
-import { useUserConfig } from "@/hooks/useUserConfig";
-import { useUserSessions } from "@/hooks/useUserSessions";
+import { useUserConfig } from "@/lib/hooks/useUserConfig";
+import { useUserSessions } from "@/lib/hooks/useUserSessions";
 import { getCapitalizedWeekday, isClassInThePast } from "@/lib/integration/common";
 import { stringifyClassPopularity } from "@/lib/popularity";
 import { ClassPopularity, RezervoClass, SessionStatus, StatusColors, UserNameWithIsSelf } from "@/types/rezervo";
