+++ src/components/modals/IntegrationUser/IntegrationUserSettingsModal.tsx
@@ -2,7 +2,7 @@ import { Dialog } from "@mui/material";
 import React, { Dispatch, SetStateAction } from "react";
 
 import IntegrationUserSettings from "@/components/modals/IntegrationUser/IntegrationUserSettings";
-import { useIntegrationUser } from "@/hooks/useIntegrationUser";
+import { useIntegrationUser } from "@/lib/hooks/useIntegrationUser";
 import { IntegrationIdentifier, IntegrationUserPayload } from "@/types/rezervo";
 
 const IntegrationUserSettingsModal = ({
