+++ src/lib/hooks/useIntegrationUser.ts
@@ -2,7 +2,7 @@ import { useUser } from "@auth0/nextjs-auth0/client";
 import useSWR from "swr";
 import useSWRMutation from "swr/mutation";
 
-import { useUserConfig } from "@/hooks/useUserConfig";
+import { useUserConfig } from "@/lib/hooks/useUserConfig";
 import { IntegrationIdentifier, IntegrationUser, IntegrationUserPayload } from "@/types/rezervo";
 import { fetcher } from "@/utils/fetchUtils";
