+++ src/lib/hooks/usePreferences.ts
