+++ src/lib/hooks/useUserConfig.ts
@@ -2,7 +2,7 @@ import { useUser } from "@auth0/nextjs-auth0/client";
 import useSWR from "swr";
 import useSWRMutation from "swr/mutation";
 
-import { useAllConfigs } from "@/hooks/useAllConfigs";
+import { useAllConfigs } from "@/lib/hooks/useAllConfigs";
 import { IntegrationConfigPayload, IntegrationIdentifier, IntegrationConfig } from "@/types/rezervo";
 import { fetcher } from "@/utils/fetchUtils";
